Fixed bug in fstring class.
The fstring class was not writing the ending null character to its
string when writing them to the file memory object. Somehow the unit
tests were not catching it(even though it should have with the final
test). Fixed bug and changed unit tests to hopefully catch that next
time.