Updated unit testing and classes.
Updated all unit testing to only use temporary files that are created
and deleted within the scope of each unit test in a tmp folder. Also
fixed two of the same bug in filemem and kincfile class when it came to
comparing their ident strings(strncmp does not work for binary data).