Modified spearman class.
Reverted old spearman opencl platform code back to original, minus the
changes needed for using the new kernel code. The kernel behaves the
same, just needs additional local space, so changes were minimal.

Also added a minimal size argument, which states the minimum number of
samples needed for a meaningful spearman coefficient, and if it is below
this threshold it returns NAN.