Modified cmatrix class.
Fixed two bugs that caused corrupted gene pair objects to be created,
with no data initialized, along with returning a null pointer when
initializing the memory for gene pair data in file memory.

Also added a default query command that dumps ALL gene correlations for
the first correlation of the first mode.