Modified spearman system.
Fixed numerous bugs in the spearman class, such as not READING the
iterators to get the actual information from file memory *sigh*. Fixed
another bug to report the time correctly. Fixed bugs in regards to
getting iterators within iterators.

Also was forced to change all work variables in the opencl code to
global instead of local... there is just too much data and not enough
local space in the GPU. This is true for the Tesla k40 too.