Added pearson analytic.
Finished coding all of pearson analytic for new ACE. I had to completely
rewrite Connor's OpenCL code because it seems he just copy pasted my
spearman code and tried to change it making a horrible frankstein
monster of code.

Pearson APPEARS to be working, or at least not crashing. However I have
no correct values to compare to so I don't know if it is giving the
right correlations.