Fixed qmake bugs.
Fixed multiple bugs in the qmake project files which made compiling with
CUDA and especially installing the software almost impossible.

The isEmpty test was being incorrectly used for the CUDADIR variable.

The program.files variable for INSTALLS was hardwiring the build
directory for its source. So if someone compiled in another shadow
directory it would fail finding anything to install.